grade view bug fixes (#2248)
* initial commit

* remove menu items for surveys

* remove more

* remove a controller

* remove manage controllers for surveys

* remove more stuff and things

* remove bad

* remove answer controller

* remove waitlist controller

* remove submitted badge

* remove unused helper

* remove usused questionnaire helper

* fix crash in waitlist.rb

* remove gemfile.lock change

* remove guard clause

* add additional checks

* remove debug

* fix issue #2090

+++ app/models/waitlist.rb
@@ -4,6 +4,8 @@ class Waitlist < ActiveRecord::Base
     unless waitlisted_topics.nil?
       waitlisted_topics.each do |waitlisted_topic|
         entry = SignedUpTeam.find_by(topic_id: waitlisted_topic.id)
+        next if entry.nil?
+        
         entry.destroy
       end
     end
